backend/requirements.txt
# TrueReact Backend Dependencies
# Google Cloud & AI
google-cloud-aiplatform>=1.38.0
google-cloud-logging>=3.8.0
google-generativeai>=0.3.0
google-adk>=0.1.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.6

# Real-time Processing
numpy>=1.26.0
scipy>=1.12.0
librosa>=0.10.1
opencv-python-headless>=4.9.0

# Async & Utilities
aiohttp>=3.9.0
asyncio>=3.4.3
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Voice Analysis
pyaudio>=0.2.14
webrtcvad>=2.0.10

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
